add changenotify.cpp
