+++ dll/win32/shell32/changenotify.cpp
@@ -1,0 +1,361 @@
+/*
+ * PROJECT:     shell32
+ * LICENSE:     LGPL-2.1-or-later (https://spdx.org/licenses/LGPL-2.1-or-later)
+ * PURPOSE:     Shell change notification
+ * COPYRIGHT:   Copyright 2020 Katayama Hirofumi MZ (katayama.hirofumi.mz@gmail.com)
+ */
+
+#include "precomp.h"
+
+WINE_DEFAULT_DEBUG_CHANNEL(shcn);
+
+CRITICAL_SECTION SHELL32_ChangenotifyCS;
+
+// This function will be called from DllMain!DLL_PROCESS_ATTACH.
+EXTERN_C void InitChangeNotifications(void)
+{
+    InitializeCriticalSection(&SHELL32_ChangenotifyCS);
+}
+
+// This function will be called from DllMain!DLL_PROCESS_DETACH.
+EXTERN_C void FreeChangeNotifications(void)
+{
+    HWND hwndWorker;
+    hwndWorker = DoGetNewDeliveryWorker();
+    SendMessageW(hwndWorker, WM_WORKER_REMOVEBYPID, GetCurrentProcessId(), 0);
+    DeleteCriticalSection(&SHELL32_ChangenotifyCS);
+}
+
+// This function handles the case of SHCNE_FREESPACE.
+static void DoNotifyFreeSpace(LPCITEMIDLIST pidl1, LPCITEMIDLIST pidl2)
+{
+    WCHAR path1[MAX_PATH], path2[MAX_PATH];
+
+    path1[0] = 0;
+    if (pidl1)
+        SHGetPathFromIDListW(pidl1, path1);
+
+    path2[0] = 0;
+    if (pidl2)
+        SHGetPathFromIDListW(pidl2, path2);
+
+    if (path1[0])
+    {
+        if (path2[0])
+            SHChangeNotify(SHCNE_FREESPACE, SHCNF_PATHW, path1, path2);
+        else
+            SHChangeNotify(SHCNE_FREESPACE, SHCNF_PATHW, path1, NULL);
+    }
+    else
+    {
+        SHChangeNotify(SHCNE_FREESPACE, SHCNF_PATHW, NULL, NULL);
+    }
+}
+
+/*************************************************************************
+ * SHChangeNotifyRegister			[SHELL32.2]
+ */
+EXTERN_C ULONG WINAPI
+SHChangeNotifyRegister(HWND hwnd, int fSources, LONG wEventMask, UINT uMsg,
+                       int cItems, SHChangeNotifyEntry *lpItems)
+{
+    HWND hwndWorker, hwndOldWorker = NULL;
+    HANDLE hShared;
+    INT iItem;
+    ULONG nRegID = INVALID_REG_ID;
+    DWORD dwOwnerPID;
+    LPREGENTRY pShare;
+
+    TRACE("(%p,0x%08x,0x%08x,0x%08x,%d,%p)\n",
+          hwnd, fSources, wEventMask, uMsg, cItems, lpItems);
+
+    // sanity check
+    if (wEventMask == 0 || cItems <= 0 || cItems > 0x7FFF || lpItems == NULL ||
+        !hwnd || !IsWindow(hwnd))
+    {
+        return INVALID_REG_ID;
+    }
+
+    // create and get the new delivery worker window
+    hwndWorker = DoGetNewDeliveryWorker();
+    if (hwndWorker == NULL)
+        return INVALID_REG_ID;
+
+    // if it is old delivery, then create the old delivery worker window
+    if ((fSources & SHCNRF_NewDelivery) == 0)
+    {
+        hwndOldWorker = hwnd = DoHireOldDeliveryWorker(hwnd, uMsg);
+        uMsg = WM_OLDDELI_HANDOVER;
+    }
+
+    if ((fSources & SHCNRF_RecursiveInterrupt) != 0 &&
+        (fSources & SHCNRF_InterruptLevel) == 0)
+    {
+        fSources &= ~SHCNRF_NewDelivery;
+    }
+
+    // The owner PID is the process ID of new delivery worker.
+    GetWindowThreadProcessId(hwndWorker, &dwOwnerPID);
+
+    EnterCriticalSection(&SHELL32_ChangenotifyCS);
+    for (iItem = 0; iItem < cItems; ++iItem)
+    {
+        // create a registration entry
+        hShared = DoCreateRegEntry(nRegID, hwnd, uMsg, fSources, wEventMask,
+                                   lpItems[iItem].fRecursive, lpItems[iItem].pidl,
+                                   dwOwnerPID, hwndOldWorker);
+        if (hShared)
+        {
+            TRACE("WM_WORKER_REGISTER: hwnd:%p, hShared:%p, pid:0x%lx\n", hwndWorker, hShared, dwOwnerPID);
+            SendMessageW(hwndWorker, WM_WORKER_REGISTER, (WPARAM)hShared, dwOwnerPID);
+
+            pShare = (LPREGENTRY)SHLockSharedEx(hShared, dwOwnerPID, FALSE);
+            if (pShare)
+            {
+                nRegID = pShare->nRegID;
+                SHUnlockShared(pShare);
+            }
+            SHFreeShared(hShared, dwOwnerPID);
+        }
+
+        if (nRegID == INVALID_REG_ID && (fSources & SHCNRF_NewDelivery) == 0)
+        {
+            ERR("Old Delivery is failed\n");
+            DestroyWindow(hwnd);
+            LeaveCriticalSection(&SHELL32_ChangenotifyCS);
+            return INVALID_REG_ID;
+        }
+    }
+    LeaveCriticalSection(&SHELL32_ChangenotifyCS);
+
+    return nRegID;
+}
+
+/*************************************************************************
+ * SHChangeNotifyDeregister			[SHELL32.4]
+ */
+EXTERN_C BOOL WINAPI
+SHChangeNotifyDeregister(ULONG hNotify)
+{
+    HWND hwndWorker;
+    LRESULT ret = 0;
+    TRACE("(0x%08x)\n", hNotify);
+
+    // get the new delivery worker window
+    hwndWorker = DoGetNewDeliveryWorker();
+    if (hwndWorker)
+    {
+        // send WM_WORKER_UNREGISTER message and try to unregister
+        ret = SendMessageW(hwndWorker, WM_WORKER_UNREGISTER, hNotify, 0);
+        if (!ret)
+        {
+            ERR("WM_WORKER_UNREGISTER failed\n");
+        }
+    }
+    return ret;
+}
+
+/*************************************************************************
+ * SHChangeNotifyUpdateEntryList       		[SHELL32.5]
+ */
+EXTERN_C BOOL WINAPI
+SHChangeNotifyUpdateEntryList(DWORD unknown1, DWORD unknown2,
+                              DWORD unknown3, DWORD unknown4)
+{
+    FIXME("(0x%08x, 0x%08x, 0x%08x, 0x%08x)\n",
+          unknown1, unknown2, unknown3, unknown4);
+
+    return TRUE;
+}
+
+/* for dumping events */
+static LPCSTR DumpEvent(LONG event)
+{
+    if (event == SHCNE_ALLEVENTS)
+        return "SHCNE_ALLEVENTS";
+#define DUMPEV(x)  ,( event & SHCNE_##x )? #x " " : ""
+    return wine_dbg_sprintf( "%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s"
+    DUMPEV(RENAMEITEM)
+    DUMPEV(CREATE)
+    DUMPEV(DELETE)
+    DUMPEV(MKDIR)
+    DUMPEV(RMDIR)
+    DUMPEV(MEDIAINSERTED)
+    DUMPEV(MEDIAREMOVED)
+    DUMPEV(DRIVEREMOVED)
+    DUMPEV(DRIVEADD)
+    DUMPEV(NETSHARE)
+    DUMPEV(NETUNSHARE)
+    DUMPEV(ATTRIBUTES)
+    DUMPEV(UPDATEDIR)
+    DUMPEV(UPDATEITEM)
+    DUMPEV(SERVERDISCONNECT)
+    DUMPEV(UPDATEIMAGE)
+    DUMPEV(DRIVEADDGUI)
+    DUMPEV(RENAMEFOLDER)
+    DUMPEV(FREESPACE)
+    DUMPEV(EXTENDED_EVENT)
+    DUMPEV(ASSOCCHANGED)
+    DUMPEV(INTERRUPT)
+    );
+#undef DUMPEV
+}
+
+/*************************************************************************
+ * SHChangeNotify				[SHELL32.@]
+ */
+EXTERN_C void WINAPI
+SHChangeNotify(LONG wEventId, UINT uFlags, LPCVOID dwItem1, LPCVOID dwItem2)
+{
+    LPITEMIDLIST pidl1 = NULL, pidl2 = NULL, pidlTemp1 = NULL, pidlTemp2 = NULL;
+    DWORD dwTick = GetTickCount();
+    WCHAR szPath1[MAX_PATH], szPath2[MAX_PATH];
+    LPWSTR psz1, psz2;
+    TRACE("(0x%08x,0x%08x,%p,%p)\n", wEventId, uFlags, dwItem1, dwItem2);
+
+    switch (uFlags & SHCNF_TYPE)
+    {
+        case SHCNF_IDLIST:
+            if (wEventId == SHCNE_FREESPACE)
+            {
+                DoNotifyFreeSpace((LPCITEMIDLIST)dwItem1, (LPCITEMIDLIST)dwItem2);
+                goto Quit;
+            }
+            pidl1 = (LPITEMIDLIST)dwItem1;
+            pidl2 = (LPITEMIDLIST)dwItem2;
+            break;
+
+        case SHCNF_PATHA:
+            psz1 = psz2 = NULL;
+            if (dwItem1)
+            {
+                MultiByteToWideChar(CP_ACP, 0, (LPCSTR)dwItem1, -1, szPath1, _countof(szPath1));
+                psz1 = szPath1;
+            }
+            if (dwItem2)
+            {
+                MultiByteToWideChar(CP_ACP, 0, (LPCSTR)dwItem2, -1, szPath2, _countof(szPath2));
+                psz2 = szPath2;
+            }
+            uFlags &= ~SHCNF_TYPE;
+            uFlags |= SHCNF_PATHW;
+            SHChangeNotify(wEventId, uFlags, psz1, psz2);
+            return;
+
+        case SHCNF_PATHW:
+            if (wEventId == SHCNE_FREESPACE)
+            {
+                /* FIXME */
+                goto Quit;
+            }
+            if (dwItem1)
+            {
+                pidl1 = pidlTemp1 = SHSimpleIDListFromPathW((LPCWSTR)dwItem1);
+            }
+            if (dwItem2)
+            {
+                pidl2 = pidlTemp2 = SHSimpleIDListFromPathW((LPCWSTR)dwItem2);
+            }
+            break;
+
+        case SHCNF_PRINTERA:
+        case SHCNF_PRINTERW:
+            FIXME("SHChangeNotify with (uFlags & SHCNF_PRINTER)\n");
+            return;
+
+        default:
+            FIXME("unknown type %08x\n", uFlags & SHCNF_TYPE);
+            return;
+    }
+
+    if (wEventId == 0 || (wEventId & SHCNE_ASSOCCHANGED) || pidl1 != NULL)
+    {
+        // transport the change
+        TRACE("notifying event %s(%x)\n", DumpEvent(wEventId), wEventId);
+        DoTransportChange(wEventId, uFlags, pidl1, pidl2, dwTick);
+    }
+
+Quit:
+    if (pidlTemp1)
+        ILFree(pidlTemp1);
+    if (pidlTemp2)
+        ILFree(pidlTemp2);
+}
+
+/*************************************************************************
+ * NTSHChangeNotifyRegister            [SHELL32.640]
+ * NOTES
+ *   Idlist is an array of structures and Count specifies how many items in the array.
+ *   count should always be one when calling SHChangeNotifyRegister, or
+ *   SHChangeNotifyDeregister will not work properly.
+ */
+EXTERN_C ULONG WINAPI
+NTSHChangeNotifyRegister(HWND hwnd, int fSources, LONG fEvents, UINT msg,
+                         int count, SHChangeNotifyEntry *idlist)
+{
+    return SHChangeNotifyRegister(hwnd, fSources | SHCNRF_NewDelivery,
+                                  fEvents, msg, count, idlist);
+}
+
+/*************************************************************************
+ * SHChangeNotification_Lock			[SHELL32.644]
+ */
+EXTERN_C HANDLE WINAPI
+SHChangeNotification_Lock(HANDLE hChange, DWORD dwProcessId, LPITEMIDLIST **lppidls,
+                          LPLONG lpwEventId)
+{
+    LPHANDBAG pHandbag;
+    TRACE("%p %08x %p %p\n", hChange, dwProcessId, lppidls, lpwEventId);
+
+    // create a handbag from delivery ticket
+    pHandbag = DoGetHandbagFromTicket(hChange, dwProcessId);
+
+    // validate the handbag
+    if (!pHandbag || pHandbag->dwMagic != HANDBAG_MAGIC)
+    {
+        ERR("pHandbag is invalid\n");
+        return NULL;
+    }
+
+    // populate the handbag
+    if (lppidls)
+        *lppidls = pHandbag->pidls;
+    if (lpwEventId)
+        *lpwEventId = pHandbag->pTicket->wEventId;
+
+    // return the handbag
+    return pHandbag;
+}
+
+/*************************************************************************
+ * SHChangeNotification_Unlock			[SHELL32.645]
+ */
+EXTERN_C BOOL WINAPI
+SHChangeNotification_Unlock(HANDLE hLock)
+{
+    LPHANDBAG pHandbag = (LPHANDBAG)hLock;
+    BOOL ret;
+    TRACE("%p\n", hLock);
+
+    // validate the handbag
+    if (!pHandbag || pHandbag->dwMagic != HANDBAG_MAGIC)
+    {
+        ERR("pHandbag is invalid\n");
+        return FALSE;
+    }
+
+    // free the handbag
+    ret = SHUnlockShared(pHandbag->pTicket);
+    LocalFree(hLock);
+    return ret;
+}
+
+/*************************************************************************
+ * NTSHChangeNotifyDeregister			[SHELL32.641]
+ */
+EXTERN_C DWORD WINAPI
+NTSHChangeNotifyDeregister(ULONG x1)
+{
+    FIXME("(0x%08x):semi stub.\n", x1);
+    return SHChangeNotifyDeregister(x1);
+}
